refactor: Move user-account route to Janus
